backend/app/requirements.txt
# Core
pytz==2018.9
Django==2.1.9
django-configurations==2.1
django-cors-headers==3.1.0
gunicorn==19.9.0

# For the persistence stores
psycopg2-binary==2.7.7

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.1
Markdown==3.0.1
django-filter==2.1.0
djangorestframework-jwt==1.11.0

# Mail
django-anymail==6.0

# Auth0
auth0-python
social-auth-app-django~=3.1
python-jose~=3.0
python-dotenv~=0.9

# Developer Tools
ipdb==0.11
ipython==7.2.0
mkdocs==1.0.4
flake8==3.7.5
django-extensions==2.2.1

# Testing
factory-boy==2.11.1
pytest==4.6.4
pytest-django==3.4.8
pytest-factoryboy==2.0.2
pytest-mock==1.10.3
coverage==4.5.2

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.93
whitenoise==4.1.3

# Error logging
sentry-sdk==0.11.1

